disable message queueing in "save matching" mode
If only matching posts are saved, there will not be 100s of messages to save per second, they could be as rare as a few messages per day. So not only we don't want to batch them by 100, but save them immediately, but also we can't use the batch save point to save the cursor periodically either. (We'll use a separate counter instead of dividing msg.seq by 100, because with Jetstream the seq doesn't increase by 1 anymore.)